# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/change-password": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/profile/image": {
      "post": {
        "operationId": "UsersController_uploadProfileImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload profile image",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteProfileImage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete profile image",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/activity-logs": {
      "get": {
        "operationId": "UsersController_getActivityLogs",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user activity logs",
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new user (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/statistics": {
      "get": {
        "operationId": "UsersController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user statistics",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/toggle-status": {
      "patch": {
        "operationId": "UsersController_toggleStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle user active/inactive status (Admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/villages": {
      "get": {
        "operationId": "VillagesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all villages"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all villages",
        "tags": [
          "villages"
        ]
      },
      "post": {
        "operationId": "VillagesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVillageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Village created successfully"
          },
          "409": {
            "description": "Village number already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new village (ADMIN only)",
        "tags": [
          "villages"
        ]
      }
    },
    "/villages/statistics": {
      "get": {
        "operationId": "VillagesController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return statistics summary"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get villages statistics",
        "tags": [
          "villages"
        ]
      }
    },
    "/villages/no/{villageNo}": {
      "get": {
        "operationId": "VillagesController_findByVillageNo",
        "parameters": [
          {
            "name": "villageNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return village by number"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get village by village number",
        "tags": [
          "villages"
        ]
      }
    },
    "/villages/{id}": {
      "get": {
        "operationId": "VillagesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return village by ID"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get village by ID",
        "tags": [
          "villages"
        ]
      },
      "patch": {
        "operationId": "VillagesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVillageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Village updated successfully"
          },
          "404": {
            "description": "Village not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update village (ADMIN only)",
        "tags": [
          "villages"
        ]
      },
      "delete": {
        "operationId": "VillagesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Village deleted successfully"
          },
          "400": {
            "description": "Cannot delete village with related data"
          },
          "404": {
            "description": "Village not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete village (ADMIN only)",
        "tags": [
          "villages"
        ]
      }
    },
    "/villages/upload/geojson/validate": {
      "post": {
        "operationId": "VillagesController_validateGeoJson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "village_boundary",
                      "district_boundary",
                      "risk_zone",
                      "infrastructure"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate GeoJSON file without saving",
        "tags": [
          "villages"
        ]
      }
    },
    "/villages/upload/geojson": {
      "post": {
        "operationId": "VillagesController_uploadGeoJson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "village_boundary",
                      "district_boundary",
                      "risk_zone",
                      "infrastructure"
                    ]
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload and save GeoJSON file",
        "tags": [
          "villages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Guardian Route API",
    "description": "\n      Guardian Route - ระบบบริหารจัดการภัยพิบัติ\n      ตำบลเวียง อำเภอฝาง จังหวัดเชียงใหม่\n      \n      ## Authentication\n      Most endpoints require JWT Bearer token authentication.\n      Use the /api/auth/login endpoint to obtain a token.\n      \n      ## Roles\n      - DEVELOPER: Full system access\n      - ADMIN: System administration\n      - SUPERVISOR: Command and control\n      - EXECUTIVE: Strategic oversight\n      - FIELD_OFFICER: Field operations\n      \n      ## Base URL\n      - Development: http://localhost:3001/api\n      - Production: https://api.guardian-route.example.com/api\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Guardian Route Team",
      "url": "https://github.com/jetci/Guardian-Route",
      "email": "support@guardian-route.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.guardian-route.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "สมชาย",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "ใจดี",
            "description": "Last name"
          },
          "phone": {
            "type": "string",
            "example": "0812345678",
            "description": "Phone number"
          },
          "department": {
            "type": "string",
            "example": "สำนักปลัด",
            "description": "Department"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "oldpassword123",
            "description": "Current password"
          },
          "newPassword": {
            "type": "string",
            "example": "newpassword123",
            "description": "New password (min 6 characters)"
          },
          "confirmPassword": {
            "type": "string",
            "example": "newpassword123",
            "description": "Confirm new password"
          }
        },
        "required": [
          "currentPassword",
          "newPassword",
          "confirmPassword"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@obtwiang.go.th"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "081-234-5678"
          },
          "department": {
            "type": "string",
            "example": "สำนักปลัด"
          },
          "role": {
            "type": "string",
            "enum": [
              "FIELD_OFFICER",
              "SUPERVISOR",
              "EXECUTIVE",
              "ADMIN",
              "DEVELOPER"
            ],
            "example": "FIELD_OFFICER"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@obtwiang.go.th"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "081-234-5678"
          },
          "department": {
            "type": "string",
            "example": "สำนักปลัด"
          },
          "role": {
            "type": "string",
            "enum": [
              "FIELD_OFFICER",
              "SUPERVISOR",
              "EXECUTIVE",
              "ADMIN",
              "DEVELOPER"
            ],
            "example": "FIELD_OFFICER"
          }
        }
      },
      "CreateVillageDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "บ้านสันทราย",
            "description": "ชื่อหมู่บ้าน"
          },
          "villageNo": {
            "type": "number",
            "example": 1,
            "description": "หมู่ที่"
          },
          "lat": {
            "type": "number",
            "example": 19.9167,
            "description": "Latitude (will be converted to centerPoint)"
          },
          "lng": {
            "type": "number",
            "example": 99.2333,
            "description": "Longitude (will be converted to centerPoint)"
          },
          "population": {
            "type": "number",
            "example": 1500,
            "description": "จำนวนประชากรรวม"
          },
          "malePopulation": {
            "type": "number",
            "example": 750,
            "description": "จำนวนประชากรชาย (will be saved as populationMale)"
          },
          "femalePopulation": {
            "type": "number",
            "example": 750,
            "description": "จำนวนประชากรหญิง (will be saved as populationFemale)"
          },
          "households": {
            "type": "number",
            "example": 350,
            "description": "จำนวนครัวเรือน"
          },
          "province": {
            "type": "string",
            "example": "เชียงใหม่",
            "description": "จังหวัด (stored in alternateNames)"
          },
          "district": {
            "type": "string",
            "example": "ฝาง",
            "description": "อำเภอ (stored in alternateNames)"
          },
          "subdistrict": {
            "type": "string",
            "example": "เวียง",
            "description": "ตำบล (stored in alternateNames)"
          }
        },
        "required": [
          "name",
          "villageNo"
        ]
      },
      "UpdateVillageDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "บ้านสันทราย",
            "description": "ชื่อหมู่บ้าน"
          },
          "villageNo": {
            "type": "number",
            "example": 1,
            "description": "หมู่ที่"
          },
          "lat": {
            "type": "number",
            "example": 19.9167,
            "description": "Latitude (will be converted to centerPoint)"
          },
          "lng": {
            "type": "number",
            "example": 99.2333,
            "description": "Longitude (will be converted to centerPoint)"
          },
          "population": {
            "type": "number",
            "example": 1500,
            "description": "จำนวนประชากรรวม"
          },
          "malePopulation": {
            "type": "number",
            "example": 750,
            "description": "จำนวนประชากรชาย (will be saved as populationMale)"
          },
          "femalePopulation": {
            "type": "number",
            "example": 750,
            "description": "จำนวนประชากรหญิง (will be saved as populationFemale)"
          },
          "households": {
            "type": "number",
            "example": 350,
            "description": "จำนวนครัวเรือน"
          },
          "province": {
            "type": "string",
            "example": "เชียงใหม่",
            "description": "จังหวัด (stored in alternateNames)"
          },
          "district": {
            "type": "string",
            "example": "ฝาง",
            "description": "อำเภอ (stored in alternateNames)"
          },
          "subdistrict": {
            "type": "string",
            "example": "เวียง",
            "description": "ตำบล (stored in alternateNames)"
          }
        }
      },
...
    }
  }
}